Fix `DataContractJsonSerializer`'s handling of -0.0
Fixes #69019.

+++ src/libraries/System.Private.DataContractSerialization/src/System/Xml/XmlConverter.cs
@@ -640,7 +640,7 @@ namespace System.Xml
             if (count == 10)
                 return false;
             if (negative)
-                result = -value;
+                result = -(float)value;
             else
                 result = value;
             return true;
@@ -692,7 +692,7 @@ namespace System.Xml
             if (count == 10)
                 return false;
             if (negative)
-                result = -value;
+                result = -(double)value;
             else
                 result = value;
             return true;
